add : menyimpan hasil rekomendasi beserta skor ke database + fix : migration

+++ database/migrations/2025_04_30_163831_create_rekomendasi_lomba_table.php
@@ -21,7 +21,7 @@ return new class extends Migration {
             $table->foreignId('lomba_id')->constrained('lomba')->onDelete('cascade');
             $table->foreignId('dosen_pembimbing_id')->nullable()->constrained('dosen')->onDelete('cascade');
             $table->enum('status', ['Disetujui', 'Ditolak', 'Pending']);
-            $table->integer('skor')->nullable();
+            $table->float('skor')->nullable();
             $table->timestamps();
         });
     }
